+++ codegen-compiler/src/main/kotlin/com/freeletics/khonshu/codegen/util/KotlinPoet.kt
@@ -121,7 +121,7 @@ internal fun Navigation?.asParameter(): ParameterSpec {
     if (this != null) {
         return ParameterSpec.builder(route.propertyName, route).build()
     }
-    return ParameterSpec.builder("intent", intent).build()
+    return ParameterSpec.builder(launchInfo.propertyName, launchInfo).build()
 }
 
 internal fun TypeName.asLambdaParameter(): TypeName {