+++ navigation/src/androidMain/kotlin/com/freeletics/khonshu/navigation/HostNavigator.kt
@@ -1,13 +1,12 @@
 package com.freeletics.khonshu.navigation
 
-import android.content.Intent
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.freeletics.khonshu.navigation.deeplinks.DeepLink
 import com.freeletics.khonshu.navigation.deeplinks.DeepLinkHandler
+import com.freeletics.khonshu.navigation.deeplinks.handleDeepLink
 import com.freeletics.khonshu.navigation.internal.InternalNavigationTestingApi
 import com.freeletics.khonshu.navigation.internal.StackEntryStoreViewModel
 import com.freeletics.khonshu.navigation.internal.StackSnapshot
@@ -18,19 +17,6 @@ import kotlinx.collections.immutable.persistentSetOf
 public abstract class HostNavigator @InternalNavigationTestingApi constructor() : Navigator {
     @InternalNavigationTestingApi
     public abstract val snapshot: State<StackSnapshot>
-
-    /**
-     * If the given [Intent] was created from a [DeepLink] or the `Uri` returned by [Intent.getData]
-     * can be handled using [deepLinkHandlers] and [deepLinkPrefixes] then the navigator will
-     * clear the current back stack and navigate to the required destinations.
-     *
-     * Returns `true` if the `Intent` contained a deeplink that was handled.
-     */
-    public abstract fun handleDeepLink(
-        intent: Intent,
-        deepLinkHandlers: ImmutableSet<DeepLinkHandler>,
-        deepLinkPrefixes: ImmutableSet<DeepLinkHandler.Prefix>,
-    ): Boolean
 
     /**
      * Allows to group multiple navigation actions and execute them atomically. The state of this [HostNavigator] will